requirements.txt
Django==3.2.7
python-telegram-bot==13.7
environs[django]==9.3.2
